requirements.txt

pycairo
libusb
pyusb
git+https://github.com/ffont/push2-python